removed all non-macro code
