--- foo.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="UTF-8"/><meta name="viewport" content="width=device-width"/><title>Will this work???</title></head><body><p>This is some content</p></body></html>
