--- public/assets/cssdsfa.css
@@ -1,1 +1,0 @@
-this is a test
