--- txt/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "hyprtxt"
-version = "0.0.1"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc_macro = true
-
-[dependencies]
-proc-macro2 = "1.0.60"
-quote = "1.0.28"
-syn = { version = "2.0.18", features = ["full", "extra-traits"]}
